refactor(api):修改Uid的生成方式, 变为8位int

+++ api/src/main/kotlin/cn/hirrao/javaee/interceptors/LoginInterceptor.kt
@@ -30,7 +30,7 @@ class LoginInterceptor @Autowired constructor(private val userService: UserServi
         try {
             val uid = parseToken(token)
             println(uid)
-            val user = userService.findByUid(uid.toLong())
+            val user = userService.findByUid(uid.toInt())
             set(user!!)
             return true
         } catch (e: Exception) {
