+++ api/src/main/kotlin/cn/hirrao/javaee/mapper/UserIntroMapper.kt
@@ -10,11 +10,11 @@ import org.apache.ibatis.annotations.Update
 @Mapper
 interface UserIntroMapper : BaseMapper<UserIntro?> {
     @Select("select * from userIntro where uid = #{uid}")
-    fun findUserIntro(uid: Long): UserIntro?
+    fun findUserIntro(uid: Int): UserIntro?
 
     @Update("update userIntro set intro=#{intro} where uid = #{uid}")
     fun updateUserIntro(uid: Long, intro: String?)
 
     @Insert("insert into userIntro(uid, intro) values(#{uid}, #{intro})")
-    fun setUserIntro(uid: Long, intro: String?)
+    fun setUserIntro(uid: Int, intro: String?)
 }
