+++ api/src/main/kotlin/cn/hirrao/javaee/service/DrugAlertService.kt
@@ -6,11 +6,11 @@ import java.time.LocalTime
 
 @Service
 interface DrugAlertService {
-    fun insertDrugAlert(alertId: Long, uid: Long, drugId: Long, alertTime: LocalTime?, eatTime: LocalDate?)
+    fun insertDrugAlert(alertId: Long, uid: Int, drugId: Long, alertTime: LocalTime?, eatTime: LocalDate?)
 
     fun updateDrugAlertIsActiveById(alertId: Long, isActive: Int)
 
     fun updateDrugAlertEatTimeById(alertId: Long, eatTime: LocalDate?)
 
-    fun deleteDrugAlertById(alertId: Long, uid: Long, drugId: Long)
+    fun deleteDrugAlertById(alertId: Long, uid: Int, drugId: Long)
 }
